# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Laivan API
  version: 0.1.0
  summary: Backend API for Laivan's (campus-scoped) accommodation marketplace.
  description: |
    Laivan is a coordination, trust, and workflow layer for fragmented campus accommodation markets.

    This contract documents the currently implemented API surface. It should describe observable HTTP behavior, not internal Go package structure.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: API server
security: []
...
paths:
...
  /v1/properties:
    get:
      operationId: listProperties
      tags:
        - Properties
      summary: List properties for a campus.
      description: |
        Returns campus-scoped properties ordered by newest first.

        This is the initial discovery feed foundation. It intentionally returns properties, not a separate generic listing model.
      parameters:
        - name: campus_id
          in: query
          required: true
          description: Campus ID that scopes marketplace discovery.
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          description: Page number. Defaults to 1.
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          required: false
          description: Number of items per page. Defaults to 20. Maximum is 100.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: sort
          in: query
          required: false
          description: Sort order. Prefix with `-` for descending. Defaults to `-created_at`.
          schema:
            type: string
            enum: [created_at, -created_at, name, -name, area, -area, lowest_price_naira, -lowest_price_naira, available_offer_count, -available_offer_count]
            default: -created_at
        - name: area
          in: query
          required: false
          description: Filter by area text (case-insensitive partial match).
          schema:
            type: string
        - name: name
          in: query
          required: false
          description: Filter by property name (case-insensitive partial match).
          schema:
            type: string
        - name: has_offers
          in: query
          required: false
          description: Filter to properties that have at least one available agent offer. Accepts `true` or `false`.
          schema:
            type: boolean
        - name: min_price
          in: query
          required: false
          description: Minimum price in naira. Filters by the property's lowest available offer price.
          schema:
            type: integer
            minimum: 1
        - name: max_price
          in: query
          required: false
          description: Maximum price in naira. Filters by the property's lowest available offer price.
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Properties returned successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertiesResponse"
              examples:
                properties:
                  value:
                    properties:
                      - id: 550e8400-e29b-41d4-a716-446655440010
                        campus_id: 550e8400-e29b-41d4-a716-446655440002
                        name: Alice Lodge
                        area: Obanla
                        landmark: Near South Gate
                        description: Gated lodge with multiple unit types near campus.
                        unit_type_count: 3
                        available_offer_count: 5
                        lowest_price_naira: 250000
                        created_at: "2026-05-01T10:00:00Z"
                        updated_at: "2026-05-01T10:00:00Z"
                    metadata:
                      current_page: 1
                      page_size: 20
                      first_page: 1
                      last_page: 1
                      total_records: 1
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      operationId: createProperty
      tags:
        - Properties
      summary: Create a property.
      description: Creates a real-world accommodation property within a campus.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePropertyRequest"
            examples:
              property:
                value:
                  campus_id: 550e8400-e29b-41d4-a716-446655440002
                  name: Alice Lodge
                  area: Obanla
                  landmark: Near South Gate
                  description: Gated lodge with multiple unit types near campus.
      responses:
        "201":
          description: Property created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/properties/{id}:
    get:
      operationId: getProperty
      tags:
        - Properties
      summary: Get a property by ID.
      parameters:
        - name: id
          in: path
          required: true
          description: Property ID.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Property returned successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyDetailResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  responses:
    BadRequest:
      description: Request body or syntax is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            badRequest:
              value:
                error:
                  code: bad_request
                  message: request body contains unknown field "unknown"
    ValidationFailed:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
          examples:
            validationFailed:
              value:
                error:
                  code: validation_failed
                  message: Request validation failed
                  fields:
                    campus_id: Campus ID is required
    NotFound:
      description: Resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            notFound:
              value:
                error:
                  code: not_found
                  message: The requested resource could not be found
...
    InternalServerError:
      description: Server encountered an unexpected problem.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            internalServerError:
              value:
                error:
                  code: internal_server_error
                  message: The server encountered a problem and could not process your request
  schemas:
...
    CreatePropertyRequest:
      type: object
      additionalProperties: false
      required:
        - campus_id
        - name
        - area
      properties:
        campus_id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 255
        area:
          type: string
          minLength: 1
          maxLength: 100
        landmark:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 1000
    Property:
      type: object
      required:
        - id
        - campus_id
        - name
        - area
        - landmark
        - description
        - unit_type_count
        - available_offer_count
        - lowest_price_naira
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        campus_id:
          type: string
          format: uuid
        name:
          type: string
        area:
          type: string
        landmark:
          type: string
          description: Approximate landmark. May be empty when unavailable.
        description:
          type: string
          description: Building-level notes. May be empty when unavailable.
        unit_type_count:
          type: integer
          minimum: 0
          description: Number of unit types at this property.
        available_offer_count:
          type: integer
          minimum: 0
          description: Number of currently available agent offers across all unit types.
        lowest_price_naira:
          type: integer
          minimum: 0
          description: Lowest price among available offers in whole Naira. 0 when no offers exist.
        thumbnail_url:
          type:
            - string
            - "null"
          format: uri
          description: Imgproxy-generated thumbnail URL for the first available image, when present.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    UnitCategory:
      type: string
      enum:
        - single_room
        - self_contained
        - room_and_parlour
        - one_bedroom_flat
        - two_bedroom_flat
        - three_bedroom_flat
        - other
    BathroomType:
      type: string
      enum:
        - private
        - shared
        - unknown
    KitchenType:
      type: string
      enum:
        - private
        - shared
        - none
        - unknown
...
    AgentOffer:
      type: object
      required:
        - id
        - property_unit_type_id
        - agent_id
        - title
        - description
        - price_naira
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        property_unit_type_id:
          type: string
          format: uuid
        agent_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          description: Agent-specific offer description. May be empty when unavailable.
        notes:
          type:
            - string
            - "null"
          maxLength: 2000
          description: Operational notes about the offer. Examples include "2 left" or "Inspection tomorrow only."
        price_naira:
          type: integer
          minimum: 1
        status:
          type: string
          enum:
            - available
            - unavailable
            - paused
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AgentSummary:
      type: object
      required:
        - id
        - display_name
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
    AgentOfferDetail:
      allOf:
        - $ref: "#/components/schemas/AgentOffer"
        - type: object
          required:
            - agent
            - media
          properties:
            agent:
              $ref: "#/components/schemas/AgentSummary"
            media:
              type: array
              items:
                $ref: "#/components/schemas/Media"
    UnitTypeDetail:
      type: object
      required:
        - id
        - property_id
        - category
        - name
        - description
        - bedroom_count
        - has_parlour
        - bathroom_type
        - kitchen_type
        - agent_offers
        - media
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        property_id:
          type: string
          format: uuid
        category:
          $ref: "#/components/schemas/UnitCategory"
        name:
          type: string
        description:
          type: string
          description: Unit type description. May be empty when unavailable.
        notes:
          type:
            - string
            - "null"
          maxLength: 2000
          description: Structural notes about the unit.
        bedroom_count:
          type:
            - integer
            - "null"
          minimum: 0
        has_parlour:
          type:
            - boolean
            - "null"
        bathroom_type:
          anyOf:
            - $ref: "#/components/schemas/BathroomType"
            - type: "null"
        kitchen_type:
          anyOf:
            - $ref: "#/components/schemas/KitchenType"
            - type: "null"
        agent_offers:
          type: array
          items:
            $ref: "#/components/schemas/AgentOfferDetail"
        media:
          type: array
          items:
            $ref: "#/components/schemas/Media"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PropertyDetail:
      type: object
      required:
        - id
        - campus_id
        - name
        - area
        - landmark
        - description
        - unit_types
        - media
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        campus_id:
          type: string
          format: uuid
        name:
          type: string
        area:
          type: string
        landmark:
          type: string
          description: Approximate landmark. May be empty when unavailable.
        description:
          type: string
          description: Building-level notes. May be empty when unavailable.
        unit_types:
          type: array
          items:
            $ref: "#/components/schemas/UnitTypeDetail"
        media:
          type: array
          items:
            $ref: "#/components/schemas/Media"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PropertyResponse:
      type: object
      required:
        - property
      properties:
        property:
          $ref: "#/components/schemas/Property"
    PropertyDetailResponse:
      type: object
      required:
        - property
      properties:
        property:
          $ref: "#/components/schemas/PropertyDetail"
...
    Media:
      type: object
      required:
        - id
        - url
        - thumbnail_url
        - medium_url
        - kind
        - size_bytes
        - created_at
      properties:
        id:
          type: string
          format: uuid
        property_id:
          type:
            - string
            - "null"
          format: uuid
        property_unit_type_id:
          type:
            - string
            - "null"
          format: uuid
        agent_offer_id:
          type:
            - string
            - "null"
          format: uuid
        uploaded_by_agent_id:
          type:
            - string
            - "null"
          format: uuid
        url:
          type: string
          format: uri
          description: Public URL for the original stored object.
        thumbnail_url:
          type: string
          format: uri
          description: Imgproxy-generated thumbnail URL.
        medium_url:
          type: string
          format: uri
          description: Imgproxy-generated medium-size URL.
        kind:
          type: string
          enum:
            - image
            - video
        caption:
          type:
            - string
            - "null"
          maxLength: 500
        content_type:
          type:
            - string
            - "null"
          examples:
            - image/jpeg
        size_bytes:
          type: integer
          minimum: 0
        created_at:
          type: string
          format: date-time
...
    PropertiesResponse:
      type: object
      required:
        - properties
        - metadata
      properties:
        properties:
          type: array
          items:
            $ref: "#/components/schemas/Property"
        metadata:
          $ref: "#/components/schemas/Metadata"
    Metadata:
      type: object
      required:
        - current_page
        - page_size
        - first_page
        - last_page
        - total_records
      properties:
        current_page:
          type: integer
        page_size:
          type: integer
        first_page:
          type: integer
        last_page:
          type: integer
        total_records:
          type: integer
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          examples:
            - bad_request
            - validation_failed
            - not_found
            - method_not_allowed
            - internal_server_error
        message:
          type: string
    ValidationError:
      allOf:
        - $ref: "#/components/schemas/Error"
        - type: object
          required:
            - fields
          properties:
            fields:
              type: object
              additionalProperties:
                type: string
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/Error"
    ValidationErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ValidationError"